refactor: split rtkQueryErrorLogger into its own module
Move the shared RTK Query error-logging middleware out of the full store so
scoped stores can reuse it without importing the full slice graph.
